refactor: customization settings
