+++ lib/features/settings/ui/appearance/appearance_screen.dart
@@ -10,11 +10,8 @@ import 'package:storii/features/settings/ui/appearance/datetime_format_tile.dart
 import 'package:storii/features/settings/ui/appearance/font_family.dart';
 import 'package:storii/features/settings/ui/appearance/font_scale.dart';
 import 'package:storii/features/settings/ui/appearance/marquee_speed_tile.dart';
-import 'package:storii/features/settings/ui/appearance/nav_label_tile.dart';
-import 'package:storii/features/settings/ui/appearance/reorder_nav_tabs.dart';
 import 'package:storii/features/settings/ui/appearance/scheme_variant_tile.dart';
 import 'package:storii/features/settings/ui/appearance/scroll_thumb_tile.dart';
-import 'package:storii/features/settings/ui/appearance/startup_nav.dart';
 import 'package:storii/features/settings/ui/appearance/system_theme_tile.dart';
 import 'package:storii/features/settings/ui/settings_header.dart';
 
@@ -108,10 +105,6 @@ class AppearanceScreen extends ConsumerWidget {
           const MarqueeSpeedTile(),
           const ScrollThumbTile(),
           const BinaryBytesTile(),
-          SettingsHeader(title: l10n.navBar),
-          const StartupNavTile(),
-          const ReorderNavTabsTile(),
-          const NavLabelTile(),
           const SizedBox(height: 100),
         ],
       ),
