+++ lib/features/settings/ui/player_settings/player_settings_screen.dart
@@ -5,7 +5,6 @@ import 'package:storii/app/config/router.dart';
 import 'package:storii/app/init.dart';
 import 'package:storii/features/settings/ui/player_settings/audio_buffer_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/fade_on_sleep_tile.dart';
-import 'package:storii/features/settings/ui/player_settings/show_seek_buttons_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/skip_tiles.dart';
 import 'package:storii/features/settings/ui/player_settings/speed_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/sync_interval_tile.dart';
@@ -43,7 +42,6 @@ class PlayerSettingsScreen extends ConsumerWidget {
       ),
       body: ListView(
         children: const [
-          ShowSeekButtonsTile(),
           SkipForwardTile(),
           SkipBackwardTile(),
           SyncIntervalTile(),
